# FW/Logging: check the message counter before fstat()ing the log file