This allows us to avoid a system call per thread, per test. However fast those system calls are (and they are fast on Linux), they add up. Here's a benchmark of the runtime of 1000 x selftest_skip in microseconds, divided by the number of threads that would have been started, with the cost of 1 of them subtracted:

| # threads  |   Before   |   After    |
| ---------: | ---------: | ---------: |
|         24 |       1.92 |       0.65 |
|         48 |       1.60 |       0.62 |
|         96 |       1.50 |       0.98 |
|        192 |       1.48 |       0.70 |

That would seem to indicate the cost of this system call is between 0.5 and 1.2 µs. The gain on Windows may be considerably bigger.

This can also be seen with selftest_pass under the same conditions, though the cost of starting the threads dwarfs the gain:

| # threads  |   Before   |   After    |
| ---------: | ---------: | ---------: |
|         24 |      37.37 |      36.94 |
|         48 |      34.06 |      33.60 |
|         96 |      32.96 |      32.56 |
|        192 |      39.78 |      38.31 |

Finally, for selftest_logs, to show that this isn't causing a meaningful regression when the test does log:

| # threads  |   Before   |   After    |
| ---------: | ---------: | ---------: |
|         24 |      50.47 |      49.85 |
|         48 |      44.95 |      44.76 |
|         96 |      42.84 |      42.94 |
|        192 |      49.12 |      49.71 |

That's between 99.7 and 101.2%.

Note: 192 threads means both sockets / NUMA domains were tested.